Initial QSSI configuration
Combined (Single) system image configuration definitions.  The
intent of this target is for generating a system image that
is common across all QTI Android targets.

This is an initial entry to set the stage and needs further
cleanup and rework.

This project contains 2 parts:
- A common definition impacting system image; for other
  targets to inherit from
- An independent QSSI system image only target for cross-
  verification

Relocation of some files from device/qcom/common at
10291a1c39346a22fc9a99aaea8b6eee28806b76.

Change-Id: I1f5f8de910c579806cf48821bfef8f2dbe1a14c0