Align QSSI system config with SDM845
Use SDM845 platform as a basis for generating the single/common
system image instead of a new "QSSI" platform path.  This allows
the least changes needed in various modules which are handled
via explicit platform name check.

This doesn't deviate from the QSSI model since SDM845 is one of
the first QSSI targets.

Change-Id: I802ee85e1d819631f47100678cfcb402a49c717c