qssi.mk: Set the various make variables for building system.img
Set the various PRODUCT_BUILD_*_IMAGE to build only the system image
(and images that have dependencies on it). Also, omit building the OTA
package. However, we still build a target files package (sans non-system
images) that can be combined with a target files package containing a
system image as a later step.

Change-Id: Ic4623b1658e0ea62955180c7d31cbdfd2fac310b