Restrict external host tools usage and export-unexport keyword
A tool must be used either as prebuilts or building them as host tools
during the build. The check was turned off by setting
TEMPORARY_DISABLE_PATH_RESTRICTIONS=true in the environment to
temporarily turn off the error checks and allow any tool to be used
(with logging).

The export and unexport keywords are obsolete, and will throw errors
when used.

Note: Not restricted for the targets having latest kernel version.

Change-Id: Ib05b1b134b0294fd7d19f219e9e393ea69b049cc